+++ packages/react-core/src/evm/contexts/thirdweb-auth.tsx
@@ -9,7 +9,6 @@ export interface ISecureStorage {
 /**
  * The configuration to use the react SDK with an [auth](https://portal.thirdweb.com/auth) server.
  *
- * @beta
  */
 export interface ThirdwebAuthConfig {
   /**